Move install, {create,ensure,clean}_venv in separate files
