+++ src/cli/commands/mod.rs
@@ -1,2 +1,8 @@
-pub mod init;
+mod develop;
+mod init;
+mod install;
+mod venv;
+pub use develop::develop;
 pub use init::init;
+pub use install::install;
+pub use venv::{clean_venv, create_venv, ensure_venv};
